+++ crates/tabby-scheduler/src/indexer.rs
